requirements.txt
Django==4.1
django-cors-headers==3.13.0
djangorestframework-camel-case==1.3.0
djangorestframework==3.13
drf_yasg==1.20.0
psycopg2==2.9.*